introduce Chocolates interface
