+++ app/src/App/Domain/Service/ChocolatesService.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace App\Domain\Service;
 
+use App\Domain\Entity\Chocolate;
 use App\Domain\Entity\User;
 use App\Domain\Service\Exception\ChocolateNotFoundException;
+use App\Domain\Value\ChocolateId;
 use App\Domain\Value\Percentage;
 use App\Domain\Value\Producer;
 use App\Domain\Value\Quantity;
 use App\Domain\Value\WrapperType;
 use App\Infrastructure\Repository\Chocolates;
-use App\Domain\Entity\Chocolate;
-use App\Domain\Value\ChocolateId;
 
 final class ChocolatesService
 {
